requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
numpy==1.23.5
scikit-learn==1.3.0
Pillow==9.5.0
scikit-image==0.20.0
joblib==1.3.0
opencv-python